Split download to separate script.
